requirements/production.txt
Django==1.9.6
Pillow
djangorestframework
